requirements.txt
openai-whisper
pydub
tkinter